Expression editor fixes
