+++ frontend/src/metabase/query_builder/components/expressions/ExpressionWidget.styled.tsx
@@ -1,9 +1,7 @@
 import styled from "@emotion/styled";
 
 import Button from "metabase/core/components/Button";
-import ExternalLink from "metabase/core/components/ExternalLink";
 import { color } from "metabase/lib/colors";
-import { Icon } from "metabase/ui";
 
 export const Container = styled.div`
   width: 472px;
@@ -27,20 +25,6 @@ export const FieldLabel = styled.label`
   letter-spacing: 0.06em;
 
   color: ${color("text-light")};
-`;
-
-export const InfoLink = styled(ExternalLink)`
-  margin-left: 4px;
-
-  &:hover,
-  :focus {
-    color: ${color("brand")};
-  }
-`;
-
-export const StyledFieldTitleIcon = styled(Icon)`
-  width: 12px;
-  height: 12px;
 `;
 
 export const Footer = styled.div`
