+++ frontend/src/metabase/query_builder/components/expressions/ExpressionWidgetInfo/ExpressionWidgetInfo.tsx
@@ -1,12 +1,15 @@
 import { t } from "ttag";
 
-import Tooltip from "metabase/core/components/Tooltip";
 import { useSelector } from "metabase/lib/redux";
 import MetabaseSettings from "metabase/lib/settings";
 import { getShowMetabaseLinks } from "metabase/selectors/whitelabel";
-import { Box } from "metabase/ui";
+import { Box, Tooltip } from "metabase/ui";
 
-import { InfoLink, StyledFieldTitleIcon } from "../ExpressionWidget.styled";
+import {
+  FieldTitleIcon,
+  InfoLink,
+  TooltipLabel,
+} from "./ExpressionWidgetInfo.styled";
 
 export const EXPRESSIONS_DOCUMENTATION_URL = MetabaseSettings.docsUrl(
   "questions/query-builder/expressions",
@@ -17,26 +20,33 @@ export function ExpressionWidgetInfo() {
 
   return showMetabaseLinks ? (
     <Tooltip
-      tooltip={t`You can reference columns here in functions or equations, like: floor([Price] - [Discount]). Click for documentation.`}
-      placement="right"
-      maxWidth={332}
+      label={
+        <TooltipLabel>
+          {t`You can reference columns here in functions or equations, like: floor([Price] - [Discount]). Click for documentation.`}
+        </TooltipLabel>
+      }
+      position="right"
     >
       <InfoLink
         target="_blank"
         href={EXPRESSIONS_DOCUMENTATION_URL}
+        tabIndex={-1}
         aria-label={t`Open expressions documentation`}
       >
-        <StyledFieldTitleIcon name="info" />
+        <FieldTitleIcon name="info" />
       </InfoLink>
     </Tooltip>
   ) : (
     <Tooltip
-      tooltip={t`You can reference columns here in functions or equations, like: floor([Price] - [Discount]).`}
-      placement="right"
-      maxWidth={332}
+      label={
+        <TooltipLabel>
+          {t`You can reference columns here in functions or equations, like: floor([Price] - [Discount]).`}
+        </TooltipLabel>
+      }
+      position="right"
     >
       <Box ml="0.25rem">
-        <StyledFieldTitleIcon name="info" />
+        <FieldTitleIcon name="info" />
       </Box>
     </Tooltip>
   );
